+++ components/CustomerBilling/CustomInput.vue
@@ -16,16 +16,6 @@ const _props = defineProps({
   editMode: { type: String, default: "UPDATE" },
 });
 const props = useDefaults(_props, "CustomerBillingCustomInput");
-
-/*****************************************************************************
- * METHODS
- *****************************************************************************/
-/**
- * props.updateProperties を使って入金予定日を未設定にする
- */
-function clearPaymentDueDate() {
-  props.updateProperties({ paymentDueDateAt: null });
-}
 </script>
 
 <template>
@@ -36,13 +26,7 @@ function clearPaymentDueDate() {
         label="入金予定日"
         :min="props.item.billingDate"
         :disabled="props.disabled || props.editMode !== 'UPDATE'"
-      />
-    </v-col>
-    <v-col cols="12">
-      <v-btn
-        text="未設定にする"
-        :disabled="props.disabled || props.editMode !== 'UPDATE'"
-        @click="clearPaymentDueDate"
+        clearable
       />
     </v-col>
   </v-row>